Update miner_data.py
# Mechanism

Some predictions that should not be removed are being deleted during `validate_historical_predictions`.  Because `existing_predictions` and `miner_predictions` contain the same entries, each wager is effectively counted twice, which causes valid predictions to exceed the daily limit and be removed.

Both `existing_predictions` and `miner_predictions` represent the predictions that the specified miner submitted on a given day.

For example, if a miner submits two predictions on the same day—one with a wager of 1 and another with a wager of 999—the second prediction is removed because the wager accumulation proceeds as follows:

### Creating the list

```
for ... in existing_predictions:
    all_predictions.append({ 
        ...,
        is_existing=True 
    })

all_predictions = [1, 999]

for ... in miner_predictions:
    all_predictions.append({ 
        ...,
        is_existing=False 
    })

all_predictions = [1, 999, 1, 999]
```

### Wager Accumulation

First, sort the list according to the `prediciton_date`.

```
all_predictions = sorted(all_predictions, ...)
all_predictions = [1, 1, 999, 999]
```

Next, accumulate the wagers and check for validity.

| Wager                    | Accumulated Total |
| ------------------------ | ----------------- |
| 1 (is_existing=True)     | 1                 |
| 1 (is_existing=False)    | 2                 |
| 999 (is_existing=True)   | *Removed*         |
| 999 (is_existing=False)  | *Removed*         |


# Solution

Remove the second `current_total += wager`, since it causes the wager to be counted twice and is not necessary.